feat: implement action decoder in step

+++ environment/utils.py
@@ -34,8 +34,8 @@ def state_to_tensor(state: State, player_color: Color) -> np.ndarray:
         featurized_state.turn_input,
         featurized_state.white_input,
         featurized_state.black_input,
-    ])
-    
+    ]).astype(np.float16)
+
 class ActionDecoder:
     """
     Decodes an action tensor produced by a DQN model into a valid Tablut action.
@@ -60,7 +60,7 @@ class ActionDecoder:
         decode(action_tensor: np.ndarray, state: State) -> Action:
             Converts a DQN-generated action tensor into a valid Tablut action.
     """
-    
+
     @staticmethod
     def _get_piece_type(action_column_index: int) -> Piece:
         """
@@ -79,7 +79,7 @@ class ActionDecoder:
         if action_column_index in range(9, 25):
             return Piece.ATTACKER
         raise IndexError("Action_column_index out of range")
-    
+
     @staticmethod
     def _num_pieces(piece: Piece) -> int:
         """
@@ -96,10 +96,10 @@ class ActionDecoder:
         if piece == Piece.DEFENDER:
             return DEFENDER_NUM
         raise ValueError("Invalid piece type")
-    
+
     @staticmethod
-    def _get_destination_coordinates(action_index: Tuple[int, int], 
-                                     moving_pawn_coords: Tuple[int, int], 
+    def _get_destination_coordinates(action_index: Tuple[int, int],
+                                     moving_pawn_coords: Tuple[int, int],
                                      state: State) -> Tuple[int, int]:
         """
         Calculate the destination coordinates for the pawn based on the action tensor and the current board state.
@@ -139,7 +139,7 @@ class ActionDecoder:
 
         # Return the move corresponding to the move index
         return valid_moves[move_index]
-    
+
     @staticmethod
     def _get_moving_pawn_coordinates(action_index: Tuple[int, int], state: State) -> Tuple[int, int]:
         """
@@ -176,7 +176,7 @@ class ActionDecoder:
             raise ValueError(f"Piece rank {piece_rank} exceeds available pieces of type {piece_type}.")
 
         return tuple(sorted_indices[piece_rank])
-    
+
     @staticmethod
     def decode(action_tensor: np.ndarray, state: State) -> Action:
         """
